refactor to separate serial and HID
